requirements.txt
python-telegram-bot==20.0
nest_asyncio
pytz
python-telegram-bot [job-queue]
requests
aiohttp
